Simplify dependencies for layout
Now not dependent on TanStack Table etc. as COLUMN_IDS is extracted into a separate lib file
